Fix int32 overflow in INT4 matmul output addressing at long prefill
Prefills of ~14k tokens or more silently produced all-zero logits.

`offs_m * stride_cm` in _int4_matmul_kernel reaches M * vocab_size. At Qwen3-8B's
151936 vocab that crosses int32's 2.15e9 ceiling once M >= ~14134: the product
wrapped negative, every masked store was dropped, and the caller got a correctly
shaped, finite, entirely zero logit tensor with no error raised. 12288 tokens
worked (1.87e9), 16384 did not (2.49e9) — exactly the observed boundary.

Row indices are now int64.

This is pre-existing and independent of the recent kernel work; it only shows up
past 14k tokens, which nothing had exercised. It is why the 16K point of the
context ladder returned garbage from the *baseline* as well as speculative
decode — worth noting, because the speculative path was the natural suspect and
was not at fault.

Also moves the paged-attention split scratch from a module-level cache to the
KV cache object. A global cache hands one engine's graph a buffer that may have
been allocated inside another engine's private CUDA-graph pool; once the first
engine was freed, replaying the second hit an illegal memory access. Ownership
now matches the graphs' lifetime, and allocation happens during warmup, outside
any capture.

Tests: 318 pass, including a regression test that genuinely allocates a
>2**31-element output (skipped when VRAM is short).

+++ engine/kernels/quant.py
@@ -179,7 +179,12 @@ def _int4_matmul_kernel(
     pid_m = tl.program_id(0)
     pid_n = tl.program_id(1)
 
-    offs_m    = pid_m * BLOCK_M + tl.arange(0, BLOCK_M)      # (BLOCK_M,)
+    # int64 row indices. tl.arange is int32, and the output offset
+    # offs_m * stride_cm reaches M * vocab_size — 16384 * 151936 = 2.49e9, past
+    # int32's 2.15e9 ceiling. The overflow wraps to a negative offset, every
+    # masked store is dropped, and the caller gets an all-zero logit tensor with
+    # no error: prefills of ~14k tokens or more silently produced garbage.
+    offs_m    = (pid_m * BLOCK_M + tl.arange(0, BLOCK_M)).to(tl.int64)   # (BLOCK_M,)
     offs_n    = pid_n * BLOCK_N + tl.arange(0, BLOCK_N)      # (BLOCK_N,)
     offs_half = tl.arange(0, GROUP_SIZE // 2)                 # (64,) — index within packed group
